refactor: 로그인 함수 util 함수로 분리
